+++ assembly/deserialize/index/typedarray.ts
@@ -1,0 +1,1 @@
+export { deserializeArrayBuffer, deserializeTypedArray } from "../simple/typedarray";