+++ assembly/serialize/index/typedarray.ts
@@ -1,0 +1,1 @@
+export { serializeArrayBufferUnsafe, serializeTypedArray } from "../simple/typedarray";